refactor(frontend): extract certificate-viewer logic into useCertificateActions hook (#777)
Move the PDF download, print, social share, copy-link and email-share
logic out of CertificateViewer into a reusable useCertificateActions
hook, leaving the component purely presentational. The Certificate
type now lives with the hook and is re-exported for compatibility.
